+++ CHANGES.rst
@@ -145,3 +145,4 @@ v1.2.4
 
 v1.2.5[unreleased]
 -----------
+* Fix issue with serializers overridden by old serializers code