+++ CHANGES.rst
@@ -149,3 +149,4 @@ v1.2.5[unreleased]
 * Implemented Availability entity with model and serializer
 * Add string formating to Availability
 * Fix string formating to Availability, using get_*_display() to resolve values
+* Implemented helpers to handle xls and csv responses