+++ CHANGES.rst
@@ -151,3 +151,4 @@ v1.2.5[unreleased]
 * Fix string formating to Availability, using get_*_display() to resolve values
 * Implemented helpers to handle xls and csv responses
 * Update pt_BR locale
+* Add image field in causes